requirements.txt
pandas
paho-mqtt
fastapi
pyjwt